Abstract account saga into its own file
